Move component to correct folder
