+++ src/components/screens/router/index.js
@@ -3,7 +3,7 @@ import { createStackNavigator, createBottomTabNavigator } from 'react-navigation
 import SignIn from '../signIn';
 import TxDetail from '../txDetail';
 import Send from '../tabs/send';
-import Bookmark from '../bookmarkPage';
+import Bookmark from '../tabs/bookmarkPage';
 import Register from '../register';
 import Wallet from '../wallet';
 import Request from '../tabs/request';
